Fragmented datum_stream.hpp
