+++ src/rdb_protocol/datum_stream.cc
@@ -5,6 +5,21 @@
 
 #include "boost_utils.hpp"
 #include "rdb_protocol/batching.hpp"
+#include "rdb_protocol/datum_stream/array.hpp"
+#include "rdb_protocol/datum_stream/eq_join.hpp"
+#include "rdb_protocol/datum_stream/fold.hpp"
+#include "rdb_protocol/datum_stream/indexed_sort.hpp"
+#include "rdb_protocol/datum_stream/lazy.hpp"
+#include "rdb_protocol/datum_stream/map.hpp"
+#include "rdb_protocol/datum_stream/offsets_of.hpp"
+#include "rdb_protocol/datum_stream/ordered_distinct.hpp"
+#include "rdb_protocol/datum_stream/ordered_union.hpp"
+#include "rdb_protocol/datum_stream/range.hpp"
+#include "rdb_protocol/datum_stream/readers.hpp"
+#include "rdb_protocol/datum_stream/readgens.hpp"
+#include "rdb_protocol/datum_stream/slice.hpp"
+#include "rdb_protocol/datum_stream/union.hpp"
+#include "rdb_protocol/datum_stream/vector.hpp"
 #include "rdb_protocol/env.hpp"
 #include "rdb_protocol/func.hpp"
 #include "rdb_protocol/geo/geojson.hpp"
@@ -17,6 +32,7 @@
 #include "rdb_protocol/geo/s2/s2latlngrect.h"
 #include "rdb_protocol/geo/s2/s2polygon.h"
 #include "rdb_protocol/geo/s2/s2polyline.h"
+#include "rdb_protocol/math_utils.hpp"
 #include "rdb_protocol/term.hpp"
 #include "rdb_protocol/val.hpp"
 #include "utils.hpp"
