+++ src/rdb_protocol/datum_stream.hpp
@@ -15,17 +15,10 @@
 #include "errors.hpp"
 #include <boost/optional.hpp>
 
-#include "concurrency/coro_pool.hpp"
-#include "concurrency/queue/unlimited_fifo.hpp"
 #include "containers/counted.hpp"
 #include "containers/scoped.hpp"
 #include "rdb_protocol/changefeed.hpp"
-#include "rdb_protocol/context.hpp"
-#include "rdb_protocol/math_utils.hpp"
-#include "rdb_protocol/order_util.hpp"
 #include "rdb_protocol/protocol.hpp"
-#include "rdb_protocol/real_table.hpp"
-#include "rdb_protocol/shards.hpp"
 
 namespace ql {
 
@@ -183,315 +176,6 @@ protected:
     const counted_t<datum_stream_t> source;
 };
 
-class offsets_of_datum_stream_t : public wrapper_datum_stream_t {
-public:
-    offsets_of_datum_stream_t(counted_t<const func_t> _f, counted_t<datum_stream_t> _source);
-
-private:
-    std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec);
-
-    counted_t<const func_t> f;
-    int64_t index;
-};
-
-class ordered_distinct_datum_stream_t : public wrapper_datum_stream_t {
-public:
-    explicit ordered_distinct_datum_stream_t(counted_t<datum_stream_t> _source);
-private:
-    std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec);
-    datum_t last_val;
-};
-
-class array_datum_stream_t : public eager_datum_stream_t {
-public:
-    array_datum_stream_t(datum_t _arr,
-                         backtrace_id_t bt);
-    virtual bool is_exhausted() const;
-    virtual feed_type_t cfeed_type() const;
-    virtual bool is_infinite() const;
-
-private:
-    virtual bool is_array() const;
-    virtual datum_t as_array(env_t *env) {
-        return is_array()
-            ? (ops_to_do() ? eager_datum_stream_t::as_array(env) : arr)
-            : datum_t();
-    }
-    virtual std::vector<datum_t>
-    next_raw_batch(env_t *env, UNUSED const batchspec_t &batchspec);
-    datum_t next(env_t *env, const batchspec_t &batchspec);
-    datum_t next_arr_el();
-
-    size_t index;
-    datum_t arr;
-};
-
-class slice_datum_stream_t : public wrapper_datum_stream_t {
-public:
-    slice_datum_stream_t(uint64_t left, uint64_t right, counted_t<datum_stream_t> src);
-private:
-    virtual std::vector<changespec_t> get_changespecs();
-    virtual std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec);
-    virtual bool is_exhausted() const;
-    virtual feed_type_t cfeed_type() const;
-    virtual bool is_infinite() const;
-    uint64_t index, left, right;
-};
-
-class indexed_sort_datum_stream_t : public wrapper_datum_stream_t {
-public:
-    indexed_sort_datum_stream_t(
-        counted_t<datum_stream_t> stream, // Must be a table with a sorting applied.
-        std::function<bool(env_t *,  // NOLINT(readability/casting)
-                           profile::sampler_t *,
-                           const datum_t &,
-                           const datum_t &)> lt_cmp);
-private:
-    virtual std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec);
-
-    std::function<bool(env_t *,  // NOLINT(readability/casting)
-                       profile::sampler_t *,
-                       const datum_t &,
-                       const datum_t &)> lt_cmp;
-    size_t index;
-    std::vector<datum_t> data;
-};
-
-struct coro_info_t;
-class coro_stream_t;
-
-class args_t;
-
-class ordered_union_datum_stream_t : public eager_datum_stream_t {
-public:
-    ordered_union_datum_stream_t(std::vector<counted_t<datum_stream_t> > &&_streams,
-                                 std::vector<
-                                 std::pair<order_direction_t,
-                                           counted_t<const func_t> > > &&_comparisons,
-                                 env_t *env,
-                                 backtrace_id_t bt);
-
-    std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec) final;
-
-    bool is_array() const final {
-        return is_array_ordered_union;
-    }
-
-    bool is_exhausted() const final;
-
-    feed_type_t cfeed_type() const final {
-        return union_type;
-    }
-
-    bool is_infinite() const final {
-        return is_infinite_ordered_union;
-    }
-
-private:
-    std::deque<counted_t<datum_stream_t> > streams;
-
-    feed_type_t union_type;
-    bool is_array_ordered_union, is_infinite_ordered_union;
-    bool is_ordered_by_field;
-
-    bool do_prelim_cache;
-
-    struct merge_cache_item_t {
-        datum_t value;
-        counted_t<datum_stream_t> source;
-    };
-
-    struct merge_less_t {
-        env_t *merge_env;
-        profile::sampler_t *merge_sampler;
-        lt_cmp_t *merge_lt_cmp;
-        bool operator()(const merge_cache_item_t &a, const merge_cache_item_t &b) {
-            // We swap `a` and `b` intentionally here, so that the
-            // `priority_queue` has the *smallest* element on top.
-            return merge_lt_cmp->operator()(merge_env,
-                                            merge_sampler,
-                                            b.value,
-                                            a.value);
-        }
-    };
-
-    lt_cmp_t lt;
-
-    std::priority_queue<merge_cache_item_t,
-                        std::vector<merge_cache_item_t>,
-                        merge_less_t> merge_cache;
-};
-
-class union_datum_stream_t : public datum_stream_t, public home_thread_mixin_t {
-public:
-    union_datum_stream_t(env_t *env,
-                         std::vector<counted_t<datum_stream_t> > &&_streams,
-                         backtrace_id_t bt,
-                         size_t expected_states = 0);
-
-    virtual void add_transformation(transform_variant_t &&tv,
-                                    backtrace_id_t bt);
-    virtual void accumulate(env_t *env, eager_acc_t *acc, const terminal_variant_t &tv);
-    virtual void accumulate_all(env_t *env, eager_acc_t *acc);
-
-    virtual bool is_array() const;
-    virtual datum_t as_array(env_t *env);
-    virtual bool is_exhausted() const;
-    virtual feed_type_t cfeed_type() const;
-    virtual bool is_infinite() const;
-
-private:
-    friend class coro_stream_t;
-
-    virtual std::vector<changespec_t> get_changespecs();
-    std::vector<datum_t >
-    next_batch_impl(env_t *env, const batchspec_t &batchspec);
-
-    // We need to keep these around to apply transformations to even though we
-    // spawn coroutines to read from them.
-    std::vector<scoped_ptr_t<coro_stream_t> > coro_streams;
-    feed_type_t union_type;
-    bool is_infinite_union;
-
-    // Set during construction.
-    scoped_ptr_t<profile::trace_t> trace;
-    scoped_ptr_t<profile::disabler_t> disabler;
-    scoped_ptr_t<env_t> coro_env;
-    // Set the first time `next_batch_impl` is called.
-    scoped_ptr_t<batchspec_t> coro_batchspec;
-
-    bool sent_init;
-    size_t ready_needed;
-
-    // A coro pool for launching reads on the individual coro_streams.
-    // If the union is not a changefeed, coro_stream_t::maybe_launch_read() is going
-    // to put reads into `read_queue` which will then be processed by `read_coro_pool`.
-    // This is to limit the degree of parallelism if a union stream is created
-    // over a large number of substreams (like in a getAll with many arguments).
-    // If the union is a changefeed, we must launch parallel reads on all streams,
-    // and this is not used (instead coro_stream_t::maybe_launch_read() will launch
-    // a coroutine directly).
-    unlimited_fifo_queue_t<std::function<void()> > read_queue;
-    calling_callback_t read_coro_callback;
-    coro_pool_t<std::function<void()> > read_coro_pool;
-
-    size_t active;
-    // We recompute this only when `next_batch_impl` returns to retain the
-    // invariant that a stream won't change from unexhausted to exhausted
-    // without attempting to read more from it.
-    bool coros_exhausted;
-    promise_t<std::exception_ptr> abort_exc;
-    scoped_ptr_t<cond_t> data_available;
-
-    std::queue<std::vector<datum_t> > queue; // FIFO
-
-    auto_drainer_t drainer;
-};
-
-class range_datum_stream_t : public eager_datum_stream_t {
-public:
-    range_datum_stream_t(bool _is_infite_range,
-                         int64_t _start,
-                         int64_t _stop,
-                         backtrace_id_t);
-
-    virtual std::vector<datum_t>
-    next_raw_batch(env_t *, const batchspec_t &batchspec);
-
-    virtual bool is_array() const {
-        return false;
-    }
-    virtual bool is_exhausted() const;
-    virtual feed_type_t cfeed_type() const {
-        return feed_type_t::not_feed;
-    }
-    virtual bool is_infinite() const {
-        return is_infinite_range;
-    }
-
-private:
-    bool is_infinite_range;
-    int64_t start, stop;
-};
-
-class map_datum_stream_t : public eager_datum_stream_t {
-public:
-    map_datum_stream_t(std::vector<counted_t<datum_stream_t> > &&_streams,
-                       counted_t<const func_t> &&_func,
-                       backtrace_id_t bt);
-
-    virtual std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec);
-
-    virtual bool is_array() const {
-        return is_array_map;
-    }
-    virtual bool is_exhausted() const;
-    virtual feed_type_t cfeed_type() const {
-        return union_type;
-    }
-    virtual bool is_infinite() const {
-        return is_infinite_map;
-    }
-
-private:
-    std::vector<counted_t<datum_stream_t> > streams;
-    counted_t<const func_t> func;
-    feed_type_t union_type;
-    bool is_array_map, is_infinite_map;
-
-    // We need to preserve this between calls because we might have gotten data
-    // from a few substreams before having to abort because we timed out on a
-    // changefeed stream and return a partial batch.  (We still time out and
-    // return partial batches, or even empty batches, for the web UI in the case
-    // of a changefeed stream.)
-    std::vector<std::deque<datum_t> > cache;
-    std::vector<datum_t> args;
-};
-
-class fold_datum_stream_t : public eager_datum_stream_t {
-public:
-    fold_datum_stream_t(counted_t<datum_stream_t> &&stream,
-                        datum_t base,
-                        counted_t<const func_t> &&_acc_func,
-                        counted_t<const func_t> &&_emit_func,
-                        counted_t<const func_t> &&_final_emit_func,
-                        backtrace_id_t bt);
-
-    std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec);
-
-    bool is_array() const final {
-        return is_array_fold;
-    }
-
-    bool is_exhausted() const final;
-
-    feed_type_t cfeed_type() const final {
-        return union_type;
-    }
-
-    bool is_infinite() const final {
-        return is_infinite_fold;
-    }
-
-private:
-    counted_t<datum_stream_t> stream;
-    counted_t<const func_t> acc_func;
-    counted_t<const func_t> emit_func;
-    counted_t<const func_t> final_emit_func;
-    feed_type_t union_type;
-    bool is_array_fold, is_infinite_fold;
-
-    datum_t acc;
-    bool do_final_emit;
-};
-
 // Every shard is in a particular state.  ACTIVE means we should read more data
 // from it, SATURATED means that there's more data to read but we didn't
 // actually use any of the data we read last time so don't bother issuing
@@ -532,531 +216,6 @@ struct active_ranges_t {
     bool totally_exhausted() const;
 };
 void debug_print(printf_buffer_t *buf, const active_ranges_t &ar);
-
-// This class generates the `read_t`s used in range reads.  It's used by
-// `reader_t` below.  Its subclasses are the different types of range reads we
-// need to do.
-class readgen_t {
-public:
-    explicit readgen_t(
-        serializable_env_t s_env,
-        std::string table_name,
-        profile_bool_t profile,
-        read_mode_t read_mode,
-        sorting_t sorting);
-    virtual ~readgen_t() { }
-
-    virtual read_t terminal_read(
-        const std::vector<transform_variant_t> &transform,
-        const terminal_variant_t &_terminal,
-        const batchspec_t &batchspec) const = 0;
-    // This has to be on `readgen_t` because we sort differently depending on
-    // the kinds of reads we're doing.
-    virtual void sindex_sort(std::vector<rget_item_t> *vec,
-                             const batchspec_t &batchspec) const = 0;
-
-    virtual read_t next_read(
-        const boost::optional<active_ranges_t> &active_ranges,
-        const boost::optional<reql_version_t> &reql_version,
-        boost::optional<changefeed_stamp_t> stamp,
-        std::vector<transform_variant_t> transform,
-        const batchspec_t &batchspec) const = 0;
-
-    virtual key_range_t original_keyrange(reql_version_t rv) const = 0;
-    virtual void restrict_active_ranges(
-        sorting_t sorting, active_ranges_t *ranges_inout) const = 0;
-    virtual boost::optional<std::string> sindex_name() const = 0;
-
-    virtual changefeed::keyspec_t::range_t get_range_spec(
-        std::vector<transform_variant_t>) const = 0;
-
-    const std::string &get_table_name() const { return table_name; }
-    read_mode_t get_read_mode() const { return read_mode; }
-    // Returns `sorting_` unless the batchspec overrides it.
-    sorting_t sorting(const batchspec_t &batchspec) const;
-protected:
-    const serializable_env_t serializable_env;
-    const std::string table_name;
-    const profile_bool_t profile;
-    const read_mode_t read_mode;
-    const sorting_t sorting_;
-};
-
-class rget_readgen_t : public readgen_t {
-public:
-    explicit rget_readgen_t(
-        serializable_env_t s_env,
-        std::string table_name,
-        const datumspec_t &datumspec,
-        profile_bool_t profile,
-        read_mode_t read_mode,
-        sorting_t sorting,
-        require_sindexes_t require_sindex_val);
-
-    virtual read_t terminal_read(
-        const std::vector<transform_variant_t> &transform,
-        const terminal_variant_t &_terminal,
-        const batchspec_t &batchspec) const;
-
-    virtual read_t next_read(
-        const boost::optional<active_ranges_t> &active_ranges,
-        const boost::optional<reql_version_t> &reql_version,
-        boost::optional<changefeed_stamp_t> stamp,
-        std::vector<transform_variant_t> transform,
-        const batchspec_t &batchspec) const;
-
-private:
-    virtual rget_read_t next_read_impl(
-        const boost::optional<active_ranges_t> &active_ranges,
-        const boost::optional<reql_version_t> &reql_version,
-        boost::optional<changefeed_stamp_t> stamp,
-        std::vector<transform_variant_t> transforms,
-        const batchspec_t &batchspec) const = 0;
-
-protected:
-    datumspec_t datumspec;
-    require_sindexes_t require_sindex_val;
-};
-
-class primary_readgen_t : public rget_readgen_t {
-public:
-    static scoped_ptr_t<readgen_t> make(
-        env_t *env,
-        std::string table_name,
-        read_mode_t read_mode,
-        const datumspec_t &datumspec = datumspec_t(datum_range_t::universe()),
-        sorting_t sorting = sorting_t::UNORDERED);
-
-private:
-    primary_readgen_t(serializable_env_t s_env,
-                      std::string table_name,
-                      const datumspec_t &datumspec,
-                      profile_bool_t profile,
-                      read_mode_t read_mode,
-                      sorting_t sorting);
-    virtual rget_read_t next_read_impl(
-        const boost::optional<active_ranges_t> &active_ranges,
-        const boost::optional<reql_version_t> &reql_version,
-        boost::optional<changefeed_stamp_t> stamp,
-        std::vector<transform_variant_t> transform,
-        const batchspec_t &batchspec) const;
-    virtual void sindex_sort(std::vector<rget_item_t> *vec,
-                             const batchspec_t &batchspec) const;
-    virtual key_range_t original_keyrange(reql_version_t rv) const;
-    virtual boost::optional<std::string> sindex_name() const;
-    void restrict_active_ranges(
-        sorting_t sorting, active_ranges_t *active_ranges_inout) const final;
-
-    virtual changefeed::keyspec_t::range_t get_range_spec(
-            std::vector<transform_variant_t> transforms) const;
-
-    boost::optional<std::map<store_key_t, uint64_t> > store_keys;
-};
-
-class sindex_readgen_t : public rget_readgen_t {
-public:
-    static scoped_ptr_t<readgen_t> make(
-        env_t *env,
-        std::string table_name,
-        read_mode_t read_mode,
-        const std::string &sindex,
-        const datumspec_t &datumspec = datumspec_t(datum_range_t::universe()),
-        sorting_t sorting = sorting_t::UNORDERED,
-        require_sindexes_t require_sindex_val = require_sindexes_t::NO);
-
-    virtual void sindex_sort(std::vector<rget_item_t> *vec,
-                             const batchspec_t &batchspec) const;
-    virtual key_range_t original_keyrange(reql_version_t rv) const;
-    virtual boost::optional<std::string> sindex_name() const;
-    void restrict_active_ranges(sorting_t, active_ranges_t *) const final { }
-private:
-    sindex_readgen_t(
-        serializable_env_t s_env,
-        std::string table_name,
-        const std::string &sindex,
-        const datumspec_t &datumspec,
-        profile_bool_t profile,
-        read_mode_t read_mode,
-        sorting_t sorting,
-        require_sindexes_t require_sindex_val);
-    virtual rget_read_t next_read_impl(
-        const boost::optional<active_ranges_t> &active_ranges,
-        const boost::optional<reql_version_t> &reql_version,
-        boost::optional<changefeed_stamp_t> stamp,
-        std::vector<transform_variant_t> transform,
-        const batchspec_t &batchspec) const;
-
-    virtual changefeed::keyspec_t::range_t get_range_spec(
-        std::vector<transform_variant_t> transforms) const;
-
-    const std::string sindex;
-    bool sent_first_read;
-};
-
-// For geospatial intersection queries
-class intersecting_readgen_t : public readgen_t {
-public:
-    static scoped_ptr_t<readgen_t> make(
-        env_t *env,
-        std::string table_name,
-        read_mode_t read_mode,
-        const std::string &sindex,
-        const datum_t &query_geometry);
-
-    virtual read_t terminal_read(
-        const std::vector<transform_variant_t> &transform,
-        const terminal_variant_t &_terminal,
-        const batchspec_t &batchspec) const;
-
-    virtual read_t next_read(
-        const boost::optional<active_ranges_t> &active_ranges,
-        const boost::optional<reql_version_t> &reql_version,
-        boost::optional<changefeed_stamp_t> stamp,
-        std::vector<transform_variant_t> transform,
-        const batchspec_t &batchspec) const;
-
-    virtual void sindex_sort(std::vector<rget_item_t> *vec,
-                             const batchspec_t &batchspec) const;
-    virtual key_range_t original_keyrange(reql_version_t rv) const;
-    virtual boost::optional<std::string> sindex_name() const;
-    void restrict_active_ranges(sorting_t, active_ranges_t *) const final { }
-
-    virtual changefeed::keyspec_t::range_t get_range_spec(
-        std::vector<transform_variant_t>) const;
-
-private:
-    intersecting_readgen_t(
-        serializable_env_t s_env,
-        std::string table_name,
-        const std::string &sindex,
-        const datum_t &query_geometry,
-        profile_bool_t profile,
-        read_mode_t read_mode);
-
-    // Analogue to rget_readgen_t::next_read_impl(), but generates an intersecting
-    // geo read.
-    intersecting_geo_read_t next_read_impl(
-        const boost::optional<active_ranges_t> &active_ranges,
-        const boost::optional<reql_version_t> &reql_version,
-        boost::optional<changefeed_stamp_t> stamp,
-        std::vector<transform_variant_t> transforms,
-        const batchspec_t &batchspec) const;
-
-    const std::string sindex;
-    const datum_t query_geometry;
-};
-
-class reader_t {
-public:
-    virtual ~reader_t() { }
-    virtual void add_transformation(transform_variant_t &&tv) = 0;
-    virtual bool add_stamp(changefeed_stamp_t stamp) = 0;
-    virtual boost::optional<active_state_t> get_active_state() = 0;
-    virtual void accumulate(env_t *env, eager_acc_t *acc,
-                            const terminal_variant_t &tv) = 0;
-    virtual void accumulate_all(env_t *env, eager_acc_t *acc) = 0;
-    virtual std::vector<datum_t> next_batch(
-        env_t *env, const batchspec_t &batchspec) = 0;
-    virtual std::vector<rget_item_t> raw_next_batch(
-        env_t *, const batchspec_t &) { unreachable(); }
-    virtual bool is_finished() const = 0;
-
-    virtual changefeed::keyspec_t get_changespec() const = 0;
-};
-
-// To handle empty range on getAll
-class empty_reader_t : public reader_t {
-public:
-    explicit empty_reader_t(counted_t<real_table_t> _table, std::string _table_name)
-      : table(std::move(_table)), table_name(std::move(_table_name)) {}
-    virtual ~empty_reader_t() {}
-    virtual void add_transformation(transform_variant_t &&) {}
-    virtual bool add_stamp(changefeed_stamp_t) {
-        r_sanity_fail();
-    }
-    virtual boost::optional<active_state_t> get_active_state() {
-        r_sanity_fail();
-    }
-    virtual void accumulate(env_t *, eager_acc_t *, const terminal_variant_t &) {}
-    virtual void accumulate_all(env_t *, eager_acc_t *) {}
-    virtual std::vector<datum_t> next_batch(env_t *, const batchspec_t &) {
-        return std::vector<datum_t>();
-    }
-    std::vector<rget_item_t> raw_next_batch(
-        env_t *, const batchspec_t &) final {
-        return std::vector<rget_item_t>{};
-    }
-    virtual bool is_finished() const {
-        return true;
-    }
-    virtual changefeed::keyspec_t get_changespec() const;
-
-private:
-    counted_t<real_table_t> table;
-    std::string table_name;
-};
-
-class vector_reader_t : public reader_t {
-public:
-    explicit vector_reader_t(std::vector<datum_t> &&_items) :
-        finished(false),
-        items(std::move(_items)) {
-    }
-    virtual ~vector_reader_t() {}
-    virtual void add_transformation(transform_variant_t &&) {
-        r_sanity_fail();
-    }
-    virtual bool add_stamp(changefeed_stamp_t) {
-        r_sanity_fail();
-    }
-    virtual boost::optional<active_state_t> get_active_state() {
-        r_sanity_fail();
-    }
-    virtual void accumulate(env_t *, eager_acc_t *, const terminal_variant_t &) {
-        r_sanity_fail();
-    }
-    virtual void accumulate_all(env_t *, eager_acc_t *) {
-        r_sanity_fail();
-    }
-    virtual std::vector<datum_t> next_batch(env_t *, const batchspec_t &) {
-        r_sanity_check(!finished);
-        finished = true;
-        return std::move(items);
-    }
-    std::vector<rget_item_t> raw_next_batch(
-        env_t *, const batchspec_t &) final {
-        r_sanity_check(!finished);
-        std::vector<rget_item_t> rget_items;
-        for (auto it = items.begin(); it != items.end(); ++it) {
-            rget_item_t item;
-            item.data = std::move(*it);
-            rget_items.push_back(std::move(item));
-        }
-        items.clear();
-        finished = true;
-        return rget_items;
-    }
-    virtual bool is_finished() const {
-        return finished;
-    }
-    virtual changefeed::keyspec_t get_changespec() const {
-        r_sanity_fail();
-    }
-
-private:
-    bool finished;
-    std::vector<datum_t> items;
-};
-
-// For reads that generate read_response_t results.
-class rget_response_reader_t : public reader_t {
-public:
-    rget_response_reader_t(
-        const counted_t<real_table_t> &table,
-        scoped_ptr_t<readgen_t> &&readgen);
-    virtual void add_transformation(transform_variant_t &&tv);
-    virtual bool add_stamp(changefeed_stamp_t stamp);
-    virtual boost::optional<active_state_t> get_active_state();
-    virtual void accumulate(env_t *env, eager_acc_t *acc, const terminal_variant_t &tv);
-    virtual void accumulate_all(env_t *env, eager_acc_t *acc) = 0;
-    virtual std::vector<datum_t> next_batch(env_t *env, const batchspec_t &batchspec);
-    virtual std::vector<rget_item_t> raw_next_batch(env_t *env,
-                                                    const batchspec_t &batchspec);
-    virtual bool is_finished() const;
-
-    virtual changefeed::keyspec_t get_changespec() const {
-        return changefeed::keyspec_t(
-            readgen->get_range_spec(transforms),
-            table,
-            readgen->get_table_name());
-    }
-
-protected:
-    raw_stream_t unshard(sorting_t sorting, rget_read_response_t &&res);
-    bool shards_exhausted() const {
-        return active_ranges ? active_ranges->totally_exhausted() : false;
-    }
-    void mark_shards_exhausted() {
-        r_sanity_check(!active_ranges);
-        active_ranges = active_ranges_t();
-    }
-
-    // Returns `true` if there's data in `items`.
-    // Overwrite this in an implementation
-    virtual bool load_items(env_t *env, const batchspec_t &batchspec) = 0;
-    rget_read_response_t do_read(env_t *env, const read_t &read);
-
-    counted_t<real_table_t> table;
-    std::vector<transform_variant_t> transforms;
-    boost::optional<changefeed_stamp_t> stamp;
-
-    bool started;
-    const scoped_ptr_t<const readgen_t> readgen;
-    boost::optional<active_ranges_t> active_ranges;
-    boost::optional<reql_version_t> reql_version;
-    std::map<uuid_u, shard_stamp_info_t> shard_stamp_infos;
-
-    // We need this to handle the SINDEX_CONSTANT case.
-    std::vector<rget_item_t> items;
-    size_t items_index;
-};
-
-class rget_reader_t : public rget_response_reader_t {
-public:
-    rget_reader_t(
-        const counted_t<real_table_t> &_table,
-        scoped_ptr_t<readgen_t> &&readgen);
-    virtual void accumulate_all(env_t *env, eager_acc_t *acc);
-
-protected:
-    // Loads new items into the `items` field of rget_response_reader_t.
-    // Returns `true` if there's data in `items`.
-    virtual bool load_items(env_t *env, const batchspec_t &batchspec);
-
-private:
-    std::vector<rget_item_t> do_range_read(env_t *env, const read_t &read);
-};
-
-// intersecting_reader_t performs filtering for duplicate documents in the stream,
-// assuming it is read in batches (otherwise that's not necessary, because the
-// shards will already provide distinct results).
-class intersecting_reader_t : public rget_response_reader_t {
-public:
-    intersecting_reader_t(
-        const counted_t<real_table_t> &_table,
-        scoped_ptr_t<readgen_t> &&readgen);
-    virtual void accumulate_all(env_t *env, eager_acc_t *acc);
-
-protected:
-    // Loads new items into the `items` field of rget_response_reader_t.
-    // Returns `true` if there's data in `items`.
-    virtual bool load_items(env_t *env, const batchspec_t &batchspec);
-
-private:
-    std::vector<rget_item_t> do_intersecting_read(env_t *env, const read_t &read);
-
-    // Each secondary index value might be inserted into a geospatial index multiple
-    // times, and we need to remove those duplicates across batches.
-    // We keep track of pairs of primary key and optional multi-index tags in order
-    // to detect and remove such duplicates.
-    std::set<std::pair<std::string, boost::optional<uint64_t> > > processed_pkey_tags;
-};
-
-class lazy_datum_stream_t;
-class eq_join_datum_stream_t : public eager_datum_stream_t {
-public:
-    eq_join_datum_stream_t(counted_t<datum_stream_t> _stream,
-                           counted_t<table_t> _table,
-                           datum_string_t _join_index,
-                           counted_t<const func_t> _predicate,
-                           bool _ordered,
-                           backtrace_id_t bt);
-
-    bool is_array() const final {
-        return is_array_eq_join;
-    }
-    bool is_infinite() const final {
-        return is_infinite_eq_join;
-    }
-    bool is_exhausted() const final;
-
-    std::vector<datum_t>
-    next_raw_batch(env_t *env, const batchspec_t &batchspec);
-
-    feed_type_t cfeed_type() const final {
-        return eq_join_type;
-    }
-
-private:
-    counted_t<datum_stream_t> stream;
-    scoped_ptr_t<reader_t> get_all_reader;
-    std::vector<rget_item_t> get_all_items;
-
-    counted_t<table_t> table;
-    datum_string_t join_index;
-
-    std::multimap<ql::datum_t,
-                  ql::datum_t> sindex_to_datum;
-
-    counted_t<const func_t> predicate;
-
-    bool ordered;
-
-    bool is_array_eq_join;
-    bool is_infinite_eq_join;
-    feed_type_t eq_join_type;
-};
-
-class lazy_datum_stream_t : public datum_stream_t {
-public:
-    lazy_datum_stream_t(
-        scoped_ptr_t<reader_t> &&_reader,
-        backtrace_id_t bt);
-
-    virtual bool is_array() const { return false; }
-    virtual datum_t as_array(UNUSED env_t *env) {
-        return datum_t();  // Cannot be converted implicitly.
-    }
-
-    bool is_exhausted() const;
-    virtual feed_type_t cfeed_type() const;
-    virtual bool is_infinite() const;
-
-    virtual bool add_stamp(changefeed_stamp_t stamp) {
-        return reader->add_stamp(std::move(stamp));
-    }
-    virtual boost::optional<active_state_t> get_active_state() {
-        return reader->get_active_state();
-    }
-
-private:
-    virtual std::vector<changespec_t> get_changespecs() {
-        return std::vector<changespec_t>{changespec_t(
-                reader->get_changespec(), counted_from_this())};
-    }
-
-    std::vector<datum_t >
-    next_batch_impl(env_t *env, const batchspec_t &batchspec);
-
-    virtual void add_transformation(transform_variant_t &&tv,
-                                    backtrace_id_t bt);
-    virtual void accumulate(env_t *env, eager_acc_t *acc, const terminal_variant_t &tv);
-    virtual void accumulate_all(env_t *env, eager_acc_t *acc);
-
-    // We use these to cache a batch so that `next` works.  There are a lot of
-    // things that are most easily written in terms of `next` that would
-    // otherwise have to do this caching themselves.
-    size_t current_batch_offset;
-    std::vector<datum_t> current_batch;
-
-    scoped_ptr_t<reader_t> reader;
-};
-
-class vector_datum_stream_t : public eager_datum_stream_t {
-public:
-    vector_datum_stream_t(
-            backtrace_id_t bt,
-            std::vector<datum_t> &&_rows,
-            boost::optional<changefeed::keyspec_t> &&_changespec);
-private:
-    datum_t next(env_t *env, const batchspec_t &bs);
-    datum_t next_impl(env_t *);
-    std::vector<datum_t> next_raw_batch(env_t *env, const batchspec_t &bs);
-
-    void add_transformation(
-        transform_variant_t &&tv, backtrace_id_t bt);
-
-    bool is_exhausted() const;
-    virtual feed_type_t cfeed_type() const;
-    bool is_array() const;
-    bool is_infinite() const;
-
-    std::vector<changespec_t> get_changespecs();
-
-    std::vector<datum_t> rows;
-    size_t index;
-    boost::optional<changefeed::keyspec_t> changespec;
-};
 
 } // namespace ql
